auctioning_platform/pyproject.toml
[tool.black]
line-length = 80
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 80

[tool.poetry]
name = "auctioning_platform"
version = "1.0.0"
authors = []
description = ""
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^21.2.0"
Flask = "^2.0.1"
injector = "^0.18.4"
SQLAlchemy = "^1.4.23"
Flask-Injector = "^0.13.0"
cattrs = "^1.8.0"
alembic = "^1.7.3"
marshmallow = "^3.13.0"
requests = "^2.26.0"
yarl = "^1.6.3"
types-requests = "^2.25.6"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
mypy = "^0.910"
pytest = "^6.2.5"
isort = "^5.9.3"
flake8 = "^3.9.2"
vcrpy = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
